backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.1.1
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
twilio==8.10.0
aiosmtplib==3.0.1
python-multipart==0.0.6
